+++ tests/util/test_string_helpers.py
@@ -9,7 +9,7 @@ import base64 as stdlib_base64
 import random
 import re
 
-from ..util.string_helpers import (
+from ...util.string_helpers import (
     UnicodeAwareBase64,
     base64,
     random_string
